app/scripts/requirements-umap.txt
numpy
scikit-learn
umap-learn
joblib
